# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FastShip API V1",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/group": {
            "get": {
                "tags": [
                    "Group Topping"
                ],
                "summary": "Get all group topping",
                "operationId": "545985736a4033004b881946c300b6ae",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "query",
                        "description": "store_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "keywords",
                        "in": "query",
                        "description": "keywords",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all group topping"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/group/create": {
            "post": {
                "tags": [
                    "Group Topping"
                ],
                "summary": "Create group topping",
                "operationId": "025d0a33f6a67dd55e5a13e7768820c1",
                "requestBody": {
                    "description": "Group topping object that needs to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name_vi": {
                                        "description": "Tên VN",
                                        "type": "string",
                                        "example": "Name vi"
                                    },
                                    "name_en": {
                                        "description": "Tên EN",
                                        "type": "string",
                                        "example": "Name en"
                                    },
                                    "name_zh": {
                                        "description": "Tên ZH",
                                        "type": "string",
                                        "example": "Name zh"
                                    },
                                    "name_hu": {
                                        "description": "Tên HU",
                                        "type": "string",
                                        "example": "name hu"
                                    },
                                    "topping_ids": {
                                        "description": "Danh sách topping liên kết",
                                        "type": "string",
                                        "example": "1,2,3"
                                    },
                                    "store_id": {
                                        "description": "ID của cửa hàng.",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create group topping Successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/group/update": {
            "post": {
                "tags": [
                    "Group Topping"
                ],
                "summary": "Update group topping",
                "operationId": "ca82f66b0cc6eea95ea3e84bfee64a76",
                "requestBody": {
                    "description": "topping object that needs to be update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "ID thể loại",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name_vi": {
                                        "description": "Tên VN",
                                        "type": "string",
                                        "example": "Name vi"
                                    },
                                    "name_en": {
                                        "description": "Tên EN",
                                        "type": "string",
                                        "example": "Name en"
                                    },
                                    "name_zh": {
                                        "description": "Tên ZH",
                                        "type": "string",
                                        "example": "Name zh"
                                    },
                                    "name_hu": {
                                        "description": "Tên HU",
                                        "type": "string",
                                        "example": "name hu"
                                    },
                                    "topping_ids": {
                                        "description": "Danh sách topping liên kết",
                                        "type": "string",
                                        "example": "1,2,3"
                                    },
                                    "store_id": {
                                        "description": "ID của cửa hàng.",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update group topping Successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/group/delete": {
            "post": {
                "tags": [
                    "Group Topping"
                ],
                "summary": "Delete group topping",
                "operationId": "2281c17b68846d73f93238bb8238fce6",
                "requestBody": {
                    "description": "Delete group topping",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Delete successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}